Implement time window for HttpStats
This makes the time_window command line arg actually have an effect.

I've tried to keep as much of the code the same for the time windowed and the
infinite window case, but in cases where the could be significant performance
improvement, I've duplicated a bit of code. I should go back and benchmark it
later.

I've also pulled in a third-party dependency for dealing with getting the local
timezone, which is necessary to correctly do time comparisons between the local
machine and the server that the log was generated on (which are likely
different machines).